Partially fix styling of modals
Partially fix the styling of modals. When I wrapped all CSS inside
`.pixel-art-react-container {...}` rules, that didn't work for the modal
since it gets attached in the DOM outside of the
`.pixel-art-react-container` div. So there are some general fixes for
this by now using Styled Components for the modal.

Specifically, we have fully fixed the keyboard shortcuts modal. There
are a couple other modals that need fixed, but we can just hide them in
the UI for now.